Refactor utilities and rename labels to tags
- Consolidate utility functions by domain (formatting, date, colors, etc)
- Rename transaction labels to tags across all files
- Unify import paths to use @/ alias consistently
- Remove duplicate utilities and organize into utils/colors/ structure
- Delete unused LabelInput component
