+++ src/components/layout/Header.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Search, Bell, Plus, Menu } from 'lucide-react';
-import Button from '../ui/Button';
-import ThemeToggle from '../ui/ThemeToggle';
+import Button from '@/components/ui/Button';
+import ThemeToggle from '@/components/ui/ThemeToggle';
 
 interface HeaderProps {
   title?: string;
